+++ lib/row_collection.dart
@@ -1,4 +1,5 @@
 library row_collection;
 
 export 'src/row_item.dart';
+export 'src/row_layout.dart';
 export 'src/separator.dart';