+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.1.0] - May 8th, 2019
+
+* **BREAKING CHANGE:** Changed `Separator.spacer` parameters to just `space`.
+* **BREAKING CHANGE:** Inside the `Separator` class, `divider` method is now `thickDivider`.
+* **BREAKING CHANGE:** Inside the `Separator` class, `thinDivider` method is now `divider`.
+* Added the `RowLayout` widget (see the example project).
+* Updated example project, to include the new features.
+* Small tweaks to `RowItem`.
+* Updated to Dart 2.2.2.
+
 ## [0.0.2] - April 28th, 2019
 
 * Small fixes to documentation